+++ engine/flink/executor/src/test/scala/pl/touk/nussknacker/engine/process/runner/FlinkTestMainSpec.scala
@@ -67,9 +67,9 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
     invocationResults("out") shouldBe
       List(ExpressionInvocationResult("proc1-id-0-1", "value", 11))
 
-    results.mockedResults("proc2") shouldBe List(MockedResult("proc1-id-0-1", "logService", "0-collectedDuringServiceInvocation"))
-    results.mockedResults("out") shouldBe List(MockedResult("proc1-id-0-1", "valueMonitor", 11))
-    results.mockedResults("eager1") shouldBe List(MockedResult("proc1-id-0-1", "collectingEager", "static-s-dynamic-0"))
+    results.externalInvocationResults("proc2") shouldBe List(ExternalInvocationResult("proc1-id-0-1", "logService", "0-collectedDuringServiceInvocation"))
+    results.externalInvocationResults("out") shouldBe List(ExternalInvocationResult("proc1-id-0-1", "valueMonitor", 11))
+    results.externalInvocationResults("eager1") shouldBe List(ExternalInvocationResult("proc1-id-0-1", "collectingEager", "static-s-dynamic-0"))
 
     MonitorEmptySink.invocationsCount.get() shouldBe 0
     LogService.invocationsCount.get() shouldBe 0
@@ -133,10 +133,10 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
         ExpressionInvocationResult("proc1-id-0-1", "value", "11 1")
       )
 
-    results.mockedResults("out") shouldBe
+    results.externalInvocationResults("out") shouldBe
       List(
-        MockedResult("proc1-id-0-0", "valueMonitor", "1 0"),
-        MockedResult("proc1-id-0-1", "valueMonitor", "11 1")
+        ExternalInvocationResult("proc1-id-0-0", "valueMonitor", "1 0"),
+        ExternalInvocationResult("proc1-id-0-1", "valueMonitor", "11 1")
       )
   }
 
@@ -249,11 +249,11 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
     val results = runFlinkTest(process, testJsonData)
 
     results.nodeResults("id") should have size 3
-    results.mockedResults("out") shouldBe
+    results.externalInvocationResults("out") shouldBe
       List(
-        MockedResult("proc1-id-0-0", "valueMonitor", SimpleJsonRecord("1", "11")),
-        MockedResult("proc1-id-0-1", "valueMonitor", SimpleJsonRecord("2", "22")),
-        MockedResult("proc1-id-0-2", "valueMonitor", SimpleJsonRecord("3", "33"))
+        ExternalInvocationResult("proc1-id-0-0", "valueMonitor", SimpleJsonRecord("1", "11")),
+        ExternalInvocationResult("proc1-id-0-1", "valueMonitor", SimpleJsonRecord("2", "22")),
+        ExternalInvocationResult("proc1-id-0-2", "valueMonitor", SimpleJsonRecord("3", "33"))
       )
   }
 
@@ -267,9 +267,9 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
     val results = runFlinkTest(process, testData)
 
     results.nodeResults("id") should have size 1
-    results.mockedResults("out") shouldBe
+    results.externalInvocationResults("out") shouldBe
       List(
-        MockedResult("proc1-id-0-0", "valueMonitor", "transformed:abc|3")
+        ExternalInvocationResult("proc1-id-0-0", "valueMonitor", "transformed:abc|3")
       )
   }
 
@@ -392,7 +392,7 @@ class FlinkTestMainSpec extends AnyFunSuite with Matchers with Inside with Befor
     val results = runFlinkTest(process, TestData.newLineSeparated(recA, recB, recC))
 
     //TODO: currently e.g. invocation results will behave strangely in this test, because we duplicate inputs and this results in duplicate context ids...
-    results.mockedResults("proc2").map(_.value.asInstanceOf[String]).sorted shouldBe List("a", "b", "c", "c").map(_ + "-collectedDuringServiceInvocation")
+    results.externalInvocationResults("proc2").map(_.value.asInstanceOf[String]).sorted shouldBe List("a", "b", "c", "c").map(_ + "-collectedDuringServiceInvocation")
   }
 
   test("should have correct run mode") {